[EDIFIKANA] #388 PaymentRecord + RentConfig BE services (#468)
* [EDIFIKANA] #387 fixing compile errors

* [EDIFIKANA] #387 Updated task models and db to remove orgId

* [EDIFIKANA] #387 Adding task api

* [EDIFIKANA] #387 updating all timestamps to use Instant. Replacing fromString to enumValueOf()

* [EDIFIKANA] #387 Creating backend task stack

* [EDIFIKANA] #387 Adding tests and related files

* [EDIFIKANA] #387 fixing some failing integ tests and updating dbs

* Addressing PR feedback

* [EDIFIKANA] HOTFIX Updating the folder organization for models and networks for easier reading.

* [EDIFIKANA] #463 updating DB to remove orgId from units, occupants, and financial

* [EDIFIKANA] #463 removed orgId from models and network models

* [EDIFIKANA] #386 Unit CRUD API — shared models, API contract, and back-end layer

Implements issue #386 (MDL-01a, API-01a, BE-01a) covering the full vertical slice
for unit management: shared library models, API operation definitions, and the
complete back-end controller/service/datastore stack.

## Shared library (edifikana/shared)
- Add `UnitModel` — client-facing domain model with all unit fields
- Add `GetUnitsQueryParams` — query params (org_id, optional property_id) for the
  list-units endpoint

## API contract (edifikana/api)
- Add `UnitApi` — defines five typed operations:
  createUnit (POST), getUnit (GET /{id}), getUnits (GET ?org_id=&property_id=),
  updateUnit (PUT /{id}), deleteUnit (DELETE /{id})

## Back-end server (edifikana/back-end)
### Domain / persistence layer
- Add `Unit` service model (mirrors DB schema, includes `createdAt`)
- Add `UnitEntity` + `CreateUnitEntity` Supabase entity with soft-delete support
- Add `UnitDatastore` interface (createUnit, getUnit, getUnits, updateUnit, deleteUnit)
- Add `SupabaseUnitDatastore` implementation:
  - Queries filter `deleted_at IS NULL` (soft-delete pattern)
  - updateUnit uses partial PATCH via `UnitEntity.CreateUnitEntity`
  - deleteUnit sets `deleted_at = clock.now()` instead of hard-deleting
- Add mapper functions in `SupabaseMappers`: `CreateUnitEntity(...)` factory and
  `UnitEntity.toUnit()` extension

### Service layer
- Add `UnitService` — thin delegation to `UnitDatastore` with `.getOrThrow()` /
  `.getOrNull()` unwrapping

### Controller layer
- Add `UnitController` with RBAC gates per endpoint:
  - createUnit → MANAGER or higher (scoped to org)
  - getUnit    → EMPLOYEE or higher (scoped to unit → org lookup)
  - getUnits   → EMPLOYEE or higher (scoped to org)
  - updateUnit → MANAGER or higher (scoped to unit → org lookup)
  - deleteUnit → ADMIN or higher (scoped to unit → org lookup)
- Add `Unit.toUnitNetworkResponse()` mapper in `NetworkMappers`

### Authorization
- Extend `RBACService` with `UnitId`-scoped `hasRole` / `hasRoleOrHigher` overloads;
  looks up the unit's `orgId` via `UnitDatastore` and delegates to the existing
  org-level role check (same pattern as DocumentId / PropertyId)
- Add `unitDatastore: UnitDatastore` constructor parameter to `RBACService`

### Dependency injection
- Register `SupabaseUnitDatastore` → `UnitDatastore` in `DatastoreModule`
- Register `UnitService` in `ServicesModule`
- Register `UnitController` as `Controller` in `ControllerModule`

## Tests
- Add `UnitControllerTest` — 10 tests covering success and auth-failure paths for
  all five CRUD operations using MockK + Koin test harness
- Add JSON fixture files for all request/response bodies
- Update `TestModule`: add `UnitController` to `TestControllerModule` and mock
  `UnitService` in `TestServiceModule`
- Fix `RBACServiceTest`: pass new `unitDatastore` mock to `RBACService` constructor


# Conflicts:
#	edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/controller/NetworkMappers.kt
#	edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/datastore/supabase/SupabaseMappers.kt
#	edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/dependencyinjection/ControllerModule.kt
#	edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/dependencyinjection/DatastoreModule.kt
#	edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/dependencyinjection/ServicesModule.kt
#	edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/service/authorization/RBACService.kt
#	edifikana/back-end/src/test/kotlin/com/cramsan/edifikana/server/dependencyinjection/TestModule.kt
#	edifikana/back-end/src/test/kotlin/com/cramsan/edifikana/server/service/authorization/RBACServiceTest.kt

* [EDIFIKANA] #386 Unit CRUD API — tests, purge, and null-update semantics

- Add UnitServiceTest covering all five CRUD operations including
  null-return and failure paths for getUnit
- Add not-found test to UnitControllerTest; fix UnitController.getUnit
  to throw NotFoundException instead of returning null (null body causes
  500 in the framework serializer)
- Add purgeUnit to UnitDatastore interface and SupabaseUnitDatastore
  implementation (hard-delete of soft-deleted records, for test cleanup)
- Add SupabaseUnitDatastoreIntegrationTest with eight tests covering
  create, get, list (full and property-filtered), update, delete, and
  not-found edge cases
- Extend SupabaseIntegrationTest base class with unitDatastore, unit
  resource tracking, createTestUnit helper, and tearDown cleanup
- Document null-means-no-change semantics on UpdateUnitNetworkRequest
  and SupabaseUnitDatastore.updateUnit


* [EDIFIKANA] Removing dependency on orgId for units

* [EDIFIKANA] Resolve test failures

* [EDIFIKANA] #388 Creating network models

* [EDIFIKANA] #388 Creating apis for payment and rent

* [EDIFIKANA] #388 Creating datastores for payment and rent

* [EDIFIKANA] #388 Creating models for payment and rent

* [EDIFIKANA] #388 Creating DS implementation for payment and rent

* [EDIFIKANA] #388 Creating DS for payment and rent

* [EDIFIKANA] #388 Creating service and controllers for payment and rent

* [EDIFIKANA] #388 Updating modules

* [EDIFIKANA] #388 Updating mappers

* [EDIFIKANA] #388 Updating test module, adding new service tests

* [EDIFIKANA] #388 Adding new controller tests

* [EDIFIKANA] #388 Updating rbac, rentConfigDS

* [EDIFIKANA] #388 creating integration tests, updating tests and services

* [EDIFIKANA] #388 updating tests, changing Long to Double for currency values,

---------

+++ edifikana/shared/src/commonMain/kotlin/com/cramsan/edifikana/lib/model/rent/RentConfigModel.kt
@@ -12,7 +12,7 @@ import com.cramsan.edifikana.lib.model.user.UserId
 data class RentConfigModel(
     val id: RentConfigId,
     val unitId: UnitId,
-    val monthlyAmount: Long,
+    val monthlyAmount: Double,
     val dueDay: Int,
     val currency: String,
     val updatedAt: Long,
